test(signup api): test

+++ src/app/api/users/route.ts
@@ -1,10 +1,30 @@
 import prisma from '@/lib/db'
+import { UserSchemaOutput } from '@/schemas/user'
 import { hashPass } from '@/utils/HashManager'
 import { NextRequest, NextResponse } from 'next/server'
 
+export type userDTO = {
+  id: string
+  firstName: string
+  lastName: string
+  email: string
+  password: string
+  createdAt: Date
+  updatedAt: Date
+}
+
 export async function GET() {
   try {
-    const users = await prisma.users.findMany()
+    const users = await prisma.users.findMany({
+      select: {
+        id: true,
+        firstName: true,
+        lastName: true,
+        email: true,
+        createdAt: true,
+        updatedAt: true,
+      },
+    })
 
     return NextResponse.json(users, { status: 200 })
   } catch (error) {
@@ -13,14 +33,12 @@ export async function GET() {
 }
 
 export async function POST(req: NextRequest) {
-  const { firstName, lastName, email, password } = await JSON.parse(JSON.stringify(req))
-
   try {
-    const findUser = await prisma.users.findFirst({ where: { email } })
+    const body = await req.json()
+    const response = UserSchemaOutput.parse(body)
+    const { firstName, lastName, email, password } = response
 
-    if (!firstName || !lastName || !email || !password) {
-      return NextResponse.json({ message: 'Missing parameter in body request' }, { status: 400 })
-    }
+    const findUser = await prisma.users.findUnique({ where: { email } })
 
     if (findUser) {
       return NextResponse.json({ message: 'E-mail already registered' }, { status: 409 })
@@ -34,6 +52,12 @@ export async function POST(req: NextRequest) {
         lastName,
         email,
         password: hashPassword,
+      },
+      select: {
+        id: true,
+        firstName: true,
+        lastName: true,
+        email: true,
       },
     })
 
@@ -49,12 +73,19 @@ export async function POST(req: NextRequest) {
   }
 }
 
-export async function PATCH(req: Request) {
+export async function PATCH(req: NextRequest) {
   const { id, email, password } = await req.json()
+
+  const findUser = await prisma.users.findUnique({ where: { id } })
+
+  if (!findUser) {
+    return NextResponse.json({ message: 'User not found' }, { status: 404 })
+  }
 
   const updatedUser = await prisma.users.update({
     where: { id },
     data: { email, password },
+    select: { id: true, firstName: true, lastName: true, email: true },
   })
 
   return NextResponse.json(
@@ -69,8 +100,20 @@ export async function PATCH(req: Request) {
 export async function DELETE(req: Request) {
   const { id } = await req.json()
 
+  const findUser = await prisma.users.findUnique({ where: { id } })
+
+  if (!findUser) {
+    return NextResponse.json({ message: 'User not found' }, { status: 404 })
+  }
+
   const deletedUser = await prisma.users.delete({
     where: { id },
+    select: {
+      id: true,
+      firstName: true,
+      lastName: true,
+      email: true,
+    },
   })
 
   return NextResponse.json(
